requirements/project.txt
Django==1.5.1
Markdown==2.3.1
Mezzanine==1.3.0
PIL==1.1.7
Pillow==2.0.0
Pygments==1.6
South==0.7.6
Werkzeug==0.8.3
argparse==1.2.1
bleach==1.2.1
boto==2.8.0
bpython==0.11
cssmin==0.1.4
distribute==0.6.24
dj-database-url==0.2.1
django-appconf==0.6
django-classy-tags==0.4
django-compressor==1.1.2
django-debug-toolbar==0.9.4
django-extensions==1.0.3
django-mptt==0.5.5
django-sekizai==0.6.1
django-storages==1.1.6
django-taggit==0.10a1
filebrowser-safe==0.2.27
geopy==0.94.2
gevent==0.13.7
grappelli-safe==0.2.21
greenlet==0.4.0
gunicorn==0.17.2
html5lib==0.95
icalendar==3.0.1b2
mezzanine-events==0.1
mezzanine-pagedown==0.2
pip-tools==0.2.1
psycopg2==2.4.6
pytz==2013b
six==1.3.0
sorl-thumbnail==11.12
wiki==0.0.19
wsgiref==0.1.2
